adding header files for exe classes
