+++ src/slam_toolbox_localization.cpp
@@ -16,39 +16,10 @@
 
 /* Author: Steven Macenski */
 
-#ifndef SLAM_TOOLBOX_SLAM_TOOLBOX_LOCALIZATION_H_
-#define SLAM_TOOLBOX_SLAM_TOOLBOX_LOCALIZATION_H_
-
-#include "slam_toolbox/slam_toolbox_common.hpp"
+#include "slam_toolbox/slam_toolbox_localization.hpp"
 
 namespace slam_toolbox
 {
-
-class LocalizationSlamToolbox : public SlamToolbox
-{
-public:
-  LocalizationSlamToolbox(ros::NodeHandle& nh);
-  ~LocalizationSlamToolbox() {};
-
-protected:
-  virtual void laserCallback(
-    const sensor_msgs::LaserScan::ConstPtr& scan) override final;
-  void localizePoseCallback(
-    const geometry_msgs::PoseWithCovarianceStampedConstPtr& msg);
-
-  virtual bool serializePoseGraphCallback(
-    slam_toolbox::SerializePoseGraph::Request& req,
-    slam_toolbox::SerializePoseGraph::Response& resp) override final;
-  virtual bool deserializePoseGraphCallback(
-    slam_toolbox::DeserializePoseGraph::Request& req,
-    slam_toolbox::DeserializePoseGraph::Response& resp) override final;
-
-  virtual bool addScan(karto::LaserRangeFinder* laser,
-    const sensor_msgs::LaserScan::ConstPtr& scan,
-    karto::Pose2& karto_pose) override final;
-
-  ros::Subscriber localization_pose_sub_;
-};
 
 /*****************************************************************************/
 LocalizationSlamToolbox::LocalizationSlamToolbox(ros::NodeHandle& nh)
@@ -252,5 +223,3 @@ int main(int argc, char** argv)
   ros::spin();
   return 0;
 }
-
-#endif //SLAM_TOOLBOX_SLAM_TOOLBOX_LOCALIZATION_H_
